requirements.txt
python-dotenv
pandas
requests
flask
numpy
sqlalchemy
pytest
pymongo[srv]==3.12
certifi
pymongo
scikit-learn
# -e .